ng/iterative-retraining/requirements.txt
scipy==1.5.0
numpy==1.13.3
torchvision==0.6.1
pandas==1.0.3
terminaltables==3.1.0
torch==1.5.1
tqdm==4.46.1
opencv_python_headless==4.2.0.34
matplotlib==3.3.0
statsmodels==0.11.1
Pillow==7.2.0
scikit_learn==0.23.2
albumentations